dir base, rand ipc, macro fix

+++ m2++_controller/boot/thread.cpp
@@ -39,7 +39,7 @@ void boot::c_thread::work()
 			}
 			catch (std::exception &e)
 			{
-				util::c_log::Instance().duo(XorStr("[ function %04x has failed => %s ]\n", &obj->func, e.what()));
+				util::c_log::Instance().duo(XorStr("[ function %04x has failed => %s ]\n"), (uint32_t)&obj->func, e.what());
 			}
 		}
 		std::this_thread::sleep_for(10ms);
